c1

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ long long int e;
 char d;
 float b;
 
-printf("Size of int: %ld bytes\n", sizeof(a));
-printf("Size of float: %ld bytes\n", sizeof(b));
-printf("Size of long long int: %ld bytes\n", sizeof(e));
-printf("Size of char: %ld bytes\n", sizeof(d));
-printf("size of long int: %1d bytes\n", sizeof(c));
+printf("Size of int: %ld bytes\n", (unsigned long)sizeof(a));
+printf("Size of float: %ld bytes\n", (unsigned long)sizeof(b));
+printf("Size of long long int: %ld bytes\n", (unsigned long)sizeof(e));
+printf("Size of char: %ld bytes\n", (unsigned long)sizeof(d));
+printf("size of long int: %1d bytes\n", (unsigned long)sizeof(c));
 return (0);
 }
